\encoding{latin1}
\name{BATCHgdwm.zero.former.change.values}
\alias{BATCHgdwm.zero.former.change.values}

\title{Sets former change value variables to zero.}

\description{Sets to zero the \emph{former.weight.change} and \emph{former.weight.bias} variables of every neuron.}

\usage{BATCHgdwm.zero.former.change.values(net)}
\arguments{
\item{net}{The neural network to modify.}
}
\value{This functions returns a modified neural network with zeroes at the former change value variables.}
\author{
Manuel Castejón Limas. 	            \email{manuel.castejon@unileon.es}\cr
Joaquin Ordieres Meré.	            \email{joaquin.ordieres@dim.unirioja.es}\cr
Ana González Marcos.                \email{ana.gonzalez@unileon.es} \cr
Alpha V. Pernía Espinoza.           \email{alpha.pernia@alum.unirioja.es}\cr
Eliseo P. Vergara Gonzalez.         \email{eliseo.vergara@dim.unirioja.es}\cr
Francisco Javier Martinez de Pisón. \email{francisco.martinez@dim.unirioja.es}\cr
Fernando Alba Elías.                \email{fernando.alba@unavarra.es}\cr
}

\references{
Simon Haykin. Neural Networks. A comprehensive foundation. 2nd Edition. \cr \cr
}

\keyword{neural}


